cleanup: custom.css does not exist
Since CKAN 2.9, web assets are bundled in the fanstatic/webassets.yml and compiled to a single file.  The reference to the file is the yaml tag given in the webassets.yml. 'custom.css' was bundled into 'styles' which is included in the 'templates/base.html' and no other references are needed since everything 'ckan_extends' everything else